refactor: rename module 'hueco' to 'verm'
Informally, this grading system is often called the "V-grade." I think
that using 'verm' would make it easier to make that connection. I also
think "verm" over "vermin" since it's shorter and is still a common in
social circles.
